+++ src/core/Entities/RegistrationFieldConfiguration.php
@@ -255,7 +255,7 @@ class RegistrationFieldConfiguration extends \MapasCulturais\Entity {
     public function getFieldName(){
         return 'field_' . $this->id;
     }
-
+    
     /**
      *
      * @return \MapasCulturais\Definitions\RegistrationFieldType
@@ -275,7 +275,7 @@ class RegistrationFieldConfiguration extends \MapasCulturais\Entity {
         'required' => $this->required,
         'fieldType' => $this->fieldType,
         'fieldOptions' => $this->fieldOptions,
-        'config' => $this->config,
+        'config' => $this->config ?: [],
         'categories' => $this->categories ?: [],
         'fieldName' => $this->getFieldName(),
         'displayOrder' => $this->displayOrder,